requirements.txt
requests==2.20.0
mysql-connector==2.1.6